requirements.txt
typer[all]
rich
google-genai
python-dotenv
duckduckgo-search
beautifulsoup4
requests
pytrends
pandas
fake-useragent
